move romc tests
